Refactor Lesson
