+++ .expo/packager-info.json
@@ -2,8 +2,8 @@
   "devToolsPort": 19002,
   "expoServerPort": 19000,
   "packagerPort": 19004,
-  "packagerPid": 87624,
-  "expoServerNgrokUrl": "https://54-yr8.anonymous.quizler-expo-ts.exp.direct",
-  "packagerNgrokUrl": "https://packager.54-yr8.anonymous.quizler-expo-ts.exp.direct",
-  "ngrokPid": 87655
+  "packagerPid": 89401,
+  "expoServerNgrokUrl": "https://54-yr8.yqbk.quizler-expo-ts.exp.direct",
+  "packagerNgrokUrl": "https://packager.54-yr8.yqbk.quizler-expo-ts.exp.direct",
+  "ngrokPid": 89417
 }
