+++ navigation/MainTabNavigator.js
@@ -6,7 +6,7 @@ import TabBarIcon from '../components/TabBarIcon';
 import HomeScreen from '../screens/HomeScreen';
 import LinksScreen from '../screens/LinksScreen';
 import SettingsScreen from '../screens/SettingsScreen';
-import LessonScreen from '../screens/Lesson';
+import LessonScreen from '../screens/Lesson/Lesson';
 
 const HomeStack = createStackNavigator({
   Home: HomeScreen,
